+++ bin/miden-cli/src/commands/new_transactions.rs
@@ -70,13 +70,12 @@ impl MintCmd {
 
         let target_account_id = parse_account_id(&client, self.target_account_id.as_str()).await?;
 
-        let transaction_request = TransactionRequestBuilder::mint_fungible_asset(
+        let transaction_request = TransactionRequestBuilder::build_mint_fungible_asset(
             fungible_asset,
             target_account_id,
             (&self.note_type).into(),
             client.rng(),
         )
-        .and_then(TransactionRequestBuilder::build)
         .map_err(|err| {
             CliError::Transaction(err.into(), "Failed to build mint transaction".to_string())
         })?;
@@ -143,13 +142,12 @@ impl SendCmd {
             target_account_id,
         );
 
-        let transaction_request = TransactionRequestBuilder::pay_to_id(
+        let transaction_request = TransactionRequestBuilder::build_pay_to_id(
             payment_transaction,
             self.recall_height.map(BlockNumber::from),
             (&self.note_type).into(),
             client.rng(),
         )
-        .and_then(TransactionRequestBuilder::build)
         .map_err(|err| {
             CliError::Transaction(err.into(), "Failed to build payment transaction".to_string())
         })?;
@@ -213,12 +211,11 @@ impl SwapCmd {
             requested_fungible_asset.into(),
         );
 
-        let transaction_request = TransactionRequestBuilder::swap(
+        let transaction_request = TransactionRequestBuilder::build_swap(
             &swap_transaction,
             (&self.note_type).into(),
             client.rng(),
         )
-        .and_then(TransactionRequestBuilder::build)
         .map_err(|err| {
             CliError::Transaction(err.into(), "Failed to build swap transaction".to_string())
         })?;