+++ crates/web-client/src/new_transactions.rs
@@ -89,13 +89,12 @@ impl WebClient {
                 JsValue::from_str("Client not initialized while generating transaction request")
             })?;
 
-            NativeTransactionRequestBuilder::mint_fungible_asset(
+            NativeTransactionRequestBuilder::build_mint_fungible_asset(
                 fungible_asset,
                 target_account_id.into(),
                 note_type.into(),
                 client.rng(),
             )
-            .and_then(NativeTransactionRequestBuilder::build)
             .map_err(|err| {
                 js_error_with_context(err, "failed to create mint transaction request")
             })?
@@ -129,13 +128,12 @@ impl WebClient {
             })?;
 
             if let Some(recall_height) = recall_height {
-                NativeTransactionRequestBuilder::pay_to_id(
+                NativeTransactionRequestBuilder::build_pay_to_id(
                     payment_transaction,
                     Some(BlockNumber::from(recall_height)),
                     note_type.into(),
                     client.rng(),
                 )
-                .and_then(NativeTransactionRequestBuilder::build)
                 .map_err(|err| {
                     js_error_with_context(
                         err,
@@ -143,13 +141,12 @@ impl WebClient {
                     )
                 })?
             } else {
-                NativeTransactionRequestBuilder::pay_to_id(
+                NativeTransactionRequestBuilder::build_pay_to_id(
                     payment_transaction,
                     None,
                     note_type.into(),
                     client.rng(),
                 )
-                .and_then(NativeTransactionRequestBuilder::build)
                 .map_err(|err| {
                     js_error_with_context(err, "failed to create send transaction request")
                 })?
@@ -175,13 +172,9 @@ impl WebClient {
                     ))
                 })?;
 
-            NativeTransactionRequestBuilder::consume_notes(native_note_ids)
-                .build()
-                .map_err(|err| {
-                    JsValue::from_str(&format!(
-                        "Failed to create Consume Transaction Request: {err}"
-                    ))
-                })?
+            NativeTransactionRequestBuilder::build_consume_notes(native_note_ids).map_err(
+                |err| js_error_with_context(err, "failed to create consume transaction request"),
+            )?
         };
 
         Ok(consume_transaction_request.into())
@@ -229,13 +222,11 @@ impl WebClient {
             .await?;
 
         if let Some(client) = self.get_mut_inner() {
-            let swap_transaction_request = NativeTransactionRequestBuilder::swap(
+            let swap_transaction_request = NativeTransactionRequestBuilder::build_swap(
                 &swap_transaction,
                 note_type.into(),
                 client.rng(),
             )
-            .map_err(|err| err.to_string())?
-            .build()
             .map_err(|err| {
                 js_error_with_context(err, "failed to create swap transaction request")
             })?;