+++ tests/integration/common.rs
@@ -328,14 +328,12 @@ pub async fn mint_note(
     // Create a Mint Tx for 1000 units of our fungible asset
     let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
     println!("Minting Asset");
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
         fungible_asset,
         basic_account_id,
         note_type,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     execute_tx_and_sync(client, fungible_asset.faucet_id(), tx_request.clone()).await;
 
@@ -354,10 +352,10 @@ pub async fn consume_notes(
     input_notes: &[InputNote],
 ) {
     println!("Consuming Note...");
-    let tx_request =
-        TransactionRequestBuilder::consume_notes(input_notes.iter().map(|n| n.id()).collect())
-            .build()
-            .unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(
+        input_notes.iter().map(|n| n.id()).collect(),
+    )
+    .unwrap();
     execute_tx_and_sync(client, account_id, tx_request).await;
 }
 
@@ -389,9 +387,8 @@ pub async fn assert_note_cannot_be_consumed_twice(
     println!("Consuming Note...");
 
     // Double-spend error expected to be received since we are consuming the same note
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_to_consume_id])
-        .build()
-        .unwrap();
+    let tx_request =
+        TransactionRequestBuilder::build_consume_notes(vec![note_to_consume_id]).unwrap();
     match client.new_transaction(consuming_account_id, tx_request).await {
         Err(ClientError::TransactionExecutorError(
             TransactionExecutorError::FetchTransactionInputsFailed(