+++ tests/integration/main.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use miden_client::{
-    ClientError, ONE,
+    ClientError, ONE, ZERO,
     account::Account,
     builder::ClientBuilder,
     note::NoteRelevance,
@@ -12,10 +12,11 @@ use miden_client::{
     },
     sync::{NoteTagSource, TX_GRACEFUL_BLOCKS},
     transaction::{
-        PaymentTransactionData, TransactionExecutorError, TransactionProver,
+        OutputNote, PaymentTransactionData, TransactionExecutorError, TransactionProver,
         TransactionProverError, TransactionRequestBuilder, TransactionStatus,
     },
 };
+use miden_lib::note::create_p2id_note;
 use miden_objects::{
     account::{AccountId, AccountStorageMode},
     asset::{Asset, FungibleAsset},
@@ -103,14 +104,12 @@ async fn test_added_notes() {
     // Mint some asset for an account not tracked by the client. It should not be stored as an
     // input note afterwards since it is not being tracked by the client
     let fungible_asset = FungibleAsset::new(faucet_account_header.id(), MINT_AMOUNT).unwrap();
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
         fungible_asset,
         AccountId::try_from(ACCOUNT_ID_REGULAR).unwrap(),
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     println!("Running Mint tx...");
     execute_tx_and_sync(&mut client, faucet_account_header.id(), tx_request).await;
@@ -140,23 +139,19 @@ async fn test_multiple_tx_on_same_block() {
 
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    let tx_request_1 = TransactionRequestBuilder::pay_to_id(
+    let tx_request_1 = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
-    let tx_request_2 = TransactionRequestBuilder::pay_to_id(
+    let tx_request_2 = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     println!("Running P2ID tx...");
@@ -234,14 +229,12 @@ async fn test_p2id_transfer() {
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2ID tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     let note = tx_request.expected_output_notes().next().unwrap().clone();
@@ -276,7 +269,7 @@ async fn test_p2id_transfer() {
 
     // Consume P2ID note
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(vec![notes[0].id()]).unwrap();
     execute_tx_and_sync(&mut client, to_account_id, tx_request).await;
 
     // Ensure we have nothing else to consume
@@ -332,14 +325,12 @@ async fn test_p2id_transfer_failing_not_enough_balance() {
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT + 1).unwrap();
     println!("Running P2ID tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         None,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     execute_failing_tx(
         &mut client,
@@ -415,14 +406,12 @@ async fn test_p2idr_transfer_consumed_by_target() {
     let current_block_num = client.get_sync_height().await.unwrap();
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num + 50),
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     execute_tx_and_sync(&mut client, from_account_id, tx_request.clone()).await;
 
@@ -434,7 +423,7 @@ async fn test_p2idr_transfer_consumed_by_target() {
     // Make the `to_account_id` consume P2IDR note
     let note_id = tx_request.expected_output_notes().next().unwrap().id();
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build().unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(vec![note_id]).unwrap();
     execute_tx_and_sync(&mut client, to_account_id, tx_request).await;
     let regular_account = client.get_account(from_account_id).await.unwrap().unwrap();
 
@@ -494,14 +483,12 @@ async fn test_p2idr_transfer_consumed_by_sender() {
     let current_block_num = client.get_sync_height().await.unwrap();
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num + 5),
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
@@ -512,7 +499,7 @@ async fn test_p2idr_transfer_consumed_by_sender() {
 
     // Check that it's still too early to consume
     println!("Consuming Note (too early)...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(vec![notes[0].id()]).unwrap();
     let transaction_execution_result = client.new_transaction(from_account_id, tx_request).await;
     assert!(transaction_execution_result.is_err_and(|err| {
         matches!(
@@ -533,7 +520,7 @@ async fn test_p2idr_transfer_consumed_by_sender() {
 
     // Consume the note with the sender account
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![notes[0].id()]).build().unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(vec![notes[0].id()]).unwrap();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
     let regular_account = client.get_account(from_account_id).await.unwrap().unwrap();
@@ -586,14 +573,12 @@ async fn test_get_consumable_notes() {
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(100.into()),
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     execute_tx_and_sync(&mut client, from_account_id, tx_request).await;
 
@@ -651,7 +636,7 @@ async fn test_get_output_notes() {
     // Do a transfer from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
     println!("Running P2ID tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(
             vec![Asset::Fungible(asset)],
             from_account_id,
@@ -661,8 +646,6 @@ async fn test_get_output_notes() {
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     let output_note_id = tx_request.expected_output_notes().next().unwrap().id();
@@ -691,14 +674,12 @@ async fn test_import_expected_notes() {
 
     wait_for_node(&mut client_2).await;
 
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
         FungibleAsset::new(faucet_account.id(), MINT_AMOUNT).unwrap(),
         client_2_account.id(),
         NoteType::Public,
         client_2.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
     client_2.sync_state().await.unwrap();
@@ -726,14 +707,12 @@ async fn test_import_expected_notes() {
     // If client 2 succesfully consumes the note, we confirm we have MMR and block header data
     consume_notes(&mut client_2, client_2_account.id(), &[input_note.try_into().unwrap()]).await;
 
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
         FungibleAsset::new(faucet_account.id(), MINT_AMOUNT).unwrap(),
         first_basic_account.id(),
         NoteType::Private,
         client_2.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
 
@@ -778,14 +757,12 @@ async fn test_import_expected_note_uncommitted() {
 
     wait_for_node(&mut client_2).await;
 
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
         FungibleAsset::new(faucet_account.id(), MINT_AMOUNT).unwrap(),
         client_2_account.id(),
         NoteType::Public,
         client_1.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
@@ -820,14 +797,12 @@ async fn test_import_expected_notes_from_the_past_as_committed() {
 
     wait_for_node(&mut client_2).await;
 
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
         FungibleAsset::new(faucet_account.id(), MINT_AMOUNT).unwrap(),
         first_basic_account.id(),
         NoteType::Public,
         client_1.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     let note: InputNoteRecord = tx_request.expected_output_notes().next().unwrap().clone().into();
 
@@ -927,14 +902,12 @@ async fn test_sync_detail_values() {
 
     // Do a transfer with recall from first account to second account
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(new_details.block_num + 5),
         NoteType::Public,
         client1.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     let note_id = tx_request.expected_output_notes().next().unwrap().id();
     execute_tx_and_sync(&mut client1, from_account_id, tx_request).await;
@@ -946,7 +919,7 @@ async fn test_sync_detail_values() {
     assert_eq!(new_details.updated_accounts.len(), 0);
 
     // Consume the note with the second account
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note_id]).build().unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(vec![note_id]).unwrap();
     execute_tx_and_sync(&mut client2, to_account_id, tx_request).await;
 
     // First client sync should have a new nullifier as the note was consumed
@@ -973,14 +946,12 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
         let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
 
         println!("Minting Asset");
-        let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+        let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
             fungible_asset,
             from_account_id,
             NoteType::Private,
             client.rng(),
         )
-        .unwrap()
-        .build()
         .unwrap();
 
         println!("Executing transaction...");
@@ -1001,14 +972,12 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
         let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
 
         println!("Minting Asset");
-        let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+        let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
             fungible_asset,
             from_account_id,
             NoteType::Private,
             client.rng(),
         )
-        .unwrap()
-        .build()
         .unwrap();
 
         println!("Executing transaction...");
@@ -1039,14 +1008,12 @@ async fn test_multiple_transactions_can_be_committed_in_different_blocks_without
         let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
 
         println!("Minting Asset");
-        let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+        let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
             fungible_asset,
             from_account_id,
             NoteType::Private,
             client.rng(),
         )
-        .unwrap()
-        .build()
         .unwrap();
 
         println!("Executing transaction...");
@@ -1152,19 +1119,17 @@ async fn test_consume_multiple_expected_notes() {
         expected_notes.partition(|note| client_notes_ids.contains(&note.id()));
 
     // Create and execute transactions
-    let tx_request_1 = TransactionRequestBuilder::consume_notes(
-        client_owned_notes.iter().map(|note| note.id()).collect(),
-    )
-    .with_authenticated_input_notes(client_owned_notes.iter().map(|note| (note.id(), None)))
-    .build()
-    .unwrap();
+    let tx_request_1 = TransactionRequestBuilder::new()
+        .with_authenticated_input_notes(client_owned_notes.iter().map(|note| (note.id(), None)))
+        .build()
+        .unwrap();
 
-    let tx_request_2 = TransactionRequestBuilder::consume_notes(
-        unauth_owned_notes.iter().map(|note| note.id()).collect(),
-    )
-    .with_unauthenticated_input_notes(unauth_owned_notes.iter().map(|note| ((*note).clone(), None)))
-    .build()
-    .unwrap();
+    let tx_request_2 = TransactionRequestBuilder::new()
+        .with_unauthenticated_input_notes(
+            unauth_owned_notes.iter().map(|note| ((*note).clone(), None)),
+        )
+        .build()
+        .unwrap();
 
     let tx_id_1 = execute_tx(&mut client, to_account_ids[0], tx_request_1).await;
     let tx_id_2 = execute_tx(&mut unauth_client, to_account_ids[1], tx_request_2).await;
@@ -1230,14 +1195,12 @@ async fn test_import_consumed_note_with_proof() {
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
 
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num),
         NoteType::Private,
         client_1.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     let note = client_1
@@ -1251,7 +1214,7 @@ async fn test_import_consumed_note_with_proof() {
     // Consume the note with the sender account
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build().unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(vec![note.id()]).unwrap();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
 
     // Import the consumed note
@@ -1290,14 +1253,12 @@ async fn test_import_consumed_note_with_id() {
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
 
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num),
         NoteType::Public,
         client_1.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     let note = client_1
@@ -1311,7 +1272,7 @@ async fn test_import_consumed_note_with_id() {
     // Consume the note with the sender account
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build().unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(vec![note.id()]).unwrap();
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
     client_2.sync_state().await.unwrap();
 
@@ -1349,14 +1310,12 @@ async fn test_discarded_transaction() {
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
 
     println!("Running P2IDR tx...");
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], from_account_id, to_account_id),
         Some(current_block_num),
         NoteType::Public,
         client_1.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     execute_tx_and_sync(&mut client_1, from_account_id, tx_request).await;
@@ -1370,7 +1329,7 @@ async fn test_discarded_transaction() {
         .clone();
 
     println!("Consuming Note...");
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![note.id()]).build().unwrap();
+    let tx_request = TransactionRequestBuilder::build_consume_notes(vec![note.id()]).unwrap();
 
     // Consume the note in client 1 but dont submit it to the node
     let tx_result = client_1.new_transaction(from_account_id, tx_request.clone()).await.unwrap();
@@ -1459,14 +1418,12 @@ async fn test_custom_transaction_prover() {
 
     let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
 
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
         fungible_asset,
         from_account_id,
         NoteType::Private,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     let transaction_execution_result =
@@ -1570,16 +1527,22 @@ async fn test_expired_transaction_fails() {
     // Create a Mint Tx for 1000 units of our fungible asset
     let fungible_asset = FungibleAsset::new(faucet_account_id, MINT_AMOUNT).unwrap();
     println!("Minting Asset");
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
-        fungible_asset,
+
+    let mint_note = create_p2id_note(
+        faucet_account_id,
         from_account_id,
+        vec![fungible_asset.into()],
         NoteType::Public,
+        ZERO,
         client.rng(),
     )
-    .unwrap()
-    .with_expiration_delta(expiration_delta)
-    .build()
     .unwrap();
+
+    let tx_request = TransactionRequestBuilder::new()
+        .with_own_output_notes(vec![OutputNote::Full(mint_note)])
+        .with_expiration_delta(expiration_delta)
+        .build()
+        .unwrap();
 
     println!("Executing transaction...");
     let transaction_execution_result =
@@ -1673,14 +1636,12 @@ async fn test_stale_transactions_discarded() {
     // Create a transaction but don't submit it to the node
     let asset = FungibleAsset::new(faucet_account_id, TRANSFER_AMOUNT).unwrap();
 
-    let tx_request = TransactionRequestBuilder::pay_to_id(
+    let tx_request = TransactionRequestBuilder::build_pay_to_id(
         PaymentTransactionData::new(vec![Asset::Fungible(asset)], account_id, account_id),
         None,
         NoteType::Public,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     // Execute the transaction but don't submit it to the node