+++ tests/integration/swap_transactions_tests.rs
@@ -92,9 +92,8 @@ async fn test_swap_fully_onchain() {
 
     println!("Consuming mint note on first client...");
 
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![account_a_mint_note_id])
-        .build()
-        .unwrap();
+    let tx_request =
+        TransactionRequestBuilder::build_consume_notes(vec![account_a_mint_note_id]).unwrap();
     execute_tx_and_sync(&mut client1, account_a.id(), tx_request).await;
 
     // Sync and consume note for accountB
@@ -109,9 +108,8 @@ async fn test_swap_fully_onchain() {
 
     println!("Consuming mint note on second client...");
 
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![account_b_mint_note_id])
-        .build()
-        .unwrap();
+    let tx_request =
+        TransactionRequestBuilder::build_consume_notes(vec![account_b_mint_note_id]).unwrap();
     execute_tx_and_sync(&mut client2, account_b.id(), tx_request).await;
 
     // Create ONCHAIN swap note (clientA offers 1 BTC in exchange of 25 ETH)
@@ -122,7 +120,7 @@ async fn test_swap_fully_onchain() {
         FungibleAsset::new(eth_faucet_account.id(), REQUESTED_ASSET_AMOUNT).unwrap();
 
     println!("Running SWAP tx...");
-    let tx_request = TransactionRequestBuilder::swap(
+    let tx_request = TransactionRequestBuilder::build_swap(
         &SwapTransactionData::new(
             account_a.id(),
             Asset::Fungible(offered_asset),
@@ -131,8 +129,6 @@ async fn test_swap_fully_onchain() {
         NoteType::Public,
         client1.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     let expected_output_notes: Vec<Note> = tx_request.expected_output_notes().cloned().collect();
@@ -161,9 +157,9 @@ async fn test_swap_fully_onchain() {
     client2.sync_state().await.unwrap();
     println!("Consuming swap note on second client...");
 
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![expected_output_notes[0].id()])
-        .build()
-        .unwrap();
+    let tx_request =
+        TransactionRequestBuilder::build_consume_notes(vec![expected_output_notes[0].id()])
+            .unwrap();
     execute_tx_and_sync(&mut client2, account_b.id(), tx_request).await;
 
     // sync on client 1, we should get the missing payback note details.
@@ -172,8 +168,7 @@ async fn test_swap_fully_onchain() {
     println!("Consuming swap payback note on first client...");
 
     let tx_request =
-        TransactionRequestBuilder::consume_notes(vec![expected_payback_note_details[0].id()])
-            .build()
+        TransactionRequestBuilder::build_consume_notes(vec![expected_payback_note_details[0].id()])
             .unwrap();
     execute_tx_and_sync(&mut client1, account_a.id(), tx_request).await;
 
@@ -313,9 +308,8 @@ async fn test_swap_private() {
 
     println!("Consuming mint note on first client...");
 
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![account_a_mint_note_id])
-        .build()
-        .unwrap();
+    let tx_request =
+        TransactionRequestBuilder::build_consume_notes(vec![account_a_mint_note_id]).unwrap();
     execute_tx_and_sync(&mut client1, account_a.id(), tx_request).await;
 
     // Sync and consume note for accountB
@@ -330,9 +324,8 @@ async fn test_swap_private() {
 
     println!("Consuming mint note on second client...");
 
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![account_b_mint_note_id])
-        .build()
-        .unwrap();
+    let tx_request =
+        TransactionRequestBuilder::build_consume_notes(vec![account_b_mint_note_id]).unwrap();
     execute_tx_and_sync(&mut client2, account_b.id(), tx_request).await;
 
     // Create ONCHAIN swap note (clientA offers 1 BTC in exchange of 25 ETH)
@@ -343,7 +336,7 @@ async fn test_swap_private() {
         FungibleAsset::new(eth_faucet_account.id(), REQUESTED_ASSET_AMOUNT).unwrap();
 
     println!("Running SWAP tx...");
-    let tx_request = TransactionRequestBuilder::swap(
+    let tx_request = TransactionRequestBuilder::build_swap(
         &SwapTransactionData::new(
             account_a.id(),
             Asset::Fungible(offered_asset),
@@ -352,8 +345,6 @@ async fn test_swap_private() {
         NoteType::Private,
         client1.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
 
     let expected_output_notes: Vec<Note> = tx_request.expected_output_notes().cloned().collect();
@@ -390,9 +381,9 @@ async fn test_swap_private() {
     // consume swap note with accountB, and check that the vault changed appropiately
     println!("Consuming swap note on second client...");
 
-    let tx_request = TransactionRequestBuilder::consume_notes(vec![expected_output_notes[0].id()])
-        .build()
-        .unwrap();
+    let tx_request =
+        TransactionRequestBuilder::build_consume_notes(vec![expected_output_notes[0].id()])
+            .unwrap();
     execute_tx_and_sync(&mut client2, account_b.id(), tx_request).await;
 
     // sync on client 1, we should get the missing payback note details.
@@ -401,8 +392,7 @@ async fn test_swap_private() {
     println!("Consuming swap payback note on first client...");
 
     let tx_request =
-        TransactionRequestBuilder::consume_notes(vec![expected_payback_note_details[0].id()])
-            .build()
+        TransactionRequestBuilder::build_consume_notes(vec![expected_payback_note_details[0].id()])
             .unwrap();
     execute_tx_and_sync(&mut client1, account_a.id(), tx_request).await;
 
@@ -480,14 +470,12 @@ async fn mint(
     let fungible_asset = FungibleAsset::new(faucet_account_id, mint_amount).unwrap();
 
     println!("Minting Asset");
-    let tx_request = TransactionRequestBuilder::mint_fungible_asset(
+    let tx_request = TransactionRequestBuilder::build_mint_fungible_asset(
         fungible_asset,
         basic_account_id,
         note_type,
         client.rng(),
     )
-    .unwrap()
-    .build()
     .unwrap();
     let id = tx_request.expected_output_notes().next().unwrap().id();
     execute_tx_and_sync(client, faucet_account_id, tx_request.clone()).await;